unify compositions
